backend/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==2.2.5
Flask-Cors==5.0.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9  # PostgreSQL adapter
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.6
MarkupSafe==2.1.5
requests==2.31.0
typing-extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
numpy
opencv-python-headless
Pillow
